+++ lib/domain/service/auth.dart
@@ -33,6 +33,7 @@ import '/domain/repository/auth.dart';
 import '/provider/drift/account.dart';
 import '/provider/drift/credentials.dart';
 import '/provider/drift/locks.dart';
+import '/provider/drift/secret.dart';
 import '/provider/gql/exceptions.dart';
 import '/routes.dart';
 import '/util/log.dart';
@@ -50,6 +51,7 @@ class AuthService extends DisposableService {
     this._credentialsProvider,
     this._accountProvider,
     this._lockProvider,
+    this._secretProvider,
   );
 
   /// Currently authorized session's [Credentials].
@@ -87,6 +89,9 @@ class AuthService extends DisposableService {
   /// [LockDriftProvider] storing the database locks.
   final LockDriftProvider _lockProvider;
 
+  /// [RefreshSecretDriftProvider] storing the [RefreshSessionSecrets].
+  final RefreshSecretDriftProvider _secretProvider;
+
   /// Authorization repository containing required authentication methods.
   final AbstractAuthRepository _authRepository;
 
@@ -120,9 +125,6 @@ class AuthService extends DisposableService {
   /// Delay between [refreshSession] invokes used to backoff it when failing
   /// over and over again.
   Duration _refreshRetryDelay = _initialRetryDelay;
-
-  /// [RefreshSessionSecrets] to use during [refreshSession].
-  RefreshSessionSecrets? _failed;
 
   /// [Stopwatch] counting since the last successful [refreshSession] occurred.
   final Map<UserId, Stopwatch> _refreshedAt = {};
@@ -680,12 +682,11 @@ class AuthService extends DisposableService {
     final bool areCurrent = userId == this.userId;
 
     Log.debug(
-      'refreshSession($userId |-> $attempt) with `isLocked` ($isLocked) and `failed` ($_failed)',
+      'refreshSession($userId |-> $attempt) with `isLocked` ($isLocked)',
       '$runtimeType',
     );
 
     LockIdentifier? dbLock;
-    RefreshSessionSecrets? secrets;
 
     try {
       // Acquire a database lock to prevent multiple refreshes of the same
@@ -909,10 +910,9 @@ class AuthService extends DisposableService {
         }
 
         try {
-          secrets = _failed ?? RefreshSessionSecrets.generate();
           final Credentials data = await _authRepository.refreshSession(
             oldCreds.refresh.secret,
-            input: secrets,
+            input: await _secretProvider.getOrCreate(oldCreds.userId),
             reconnect: areCurrent,
           );
 
@@ -921,7 +921,7 @@ class AuthService extends DisposableService {
             '$runtimeType',
           );
 
-          _failed = null;
+          await _secretProvider.delete(oldCreds.userId);
           _refreshedAt[userId]?.stop();
           _refreshedAt[userId] = Stopwatch()..start();
 
@@ -956,6 +956,7 @@ class AuthService extends DisposableService {
             // Remove stale [Credentials].
             accounts.remove(oldCreds.userId);
             await _credentialsProvider.delete(oldCreds.userId);
+            await _secretProvider.delete(oldCreds.userId);
           }
 
           _refreshRetryDelay = _initialRetryDelay;
@@ -966,7 +967,6 @@ class AuthService extends DisposableService {
       await _lockProvider.release(dbLock);
     } on RefreshSessionException catch (_) {
       _refreshRetryDelay = _initialRetryDelay;
-      _failed = null;
 
       if (dbLock != null) {
         await _lockProvider.release(dbLock);
@@ -978,8 +978,6 @@ class AuthService extends DisposableService {
         'refreshSession($userId |-> $attempt): ⛔️ exception occurred: $e',
         '$runtimeType',
       );
-
-      _failed = secrets;
 
       if (dbLock != null) {
         await _lockProvider.release(dbLock);
@@ -1073,6 +1071,7 @@ class AuthService extends DisposableService {
     final UserId? id = userId;
     if (id != null) {
       _credentialsProvider.delete(id);
+      _secretProvider.delete(id);
       _refreshTimers.remove(id)?.cancel();
       accounts.remove(id);
       WebUtils.removeCredentials(id);