+++ lib/store/chat.dart
@@ -1117,7 +1117,7 @@ class ChatRepository extends DisposableInterface
           break;
 
         case EditChatMessageErrorCode.notAuthor:
-        case EditChatMessageErrorCode.noTextAndNoAttachment:
+        case EditChatMessageErrorCode.noContent:
           // No-op.
           break;
       }
@@ -1165,7 +1165,7 @@ class ChatRepository extends DisposableInterface
           switch (e.code) {
             case DeleteChatMessageErrorCode.notAuthor:
             case DeleteChatMessageErrorCode.quoted:
-            case DeleteChatMessageErrorCode.read:
+            case DeleteChatMessageErrorCode.uneditable:
               rethrow;
 
             case DeleteChatMessageErrorCode.unknownChatItem:
@@ -1226,7 +1226,7 @@ class ChatRepository extends DisposableInterface
             case DeleteChatForwardErrorCode.artemisUnknown:
             case DeleteChatForwardErrorCode.notAuthor:
             case DeleteChatForwardErrorCode.quoted:
-            case DeleteChatForwardErrorCode.read:
+            case DeleteChatForwardErrorCode.uneditable:
               rethrow;
 
             case DeleteChatForwardErrorCode.unknownChatItem:
@@ -1474,7 +1474,10 @@ class ChatRepository extends DisposableInterface
     } on ForwardChatItemsException catch (e) {
       switch (e.code) {
         case ForwardChatItemsErrorCode.blocked:
-        case ForwardChatItemsErrorCode.noTextAndNoAttachment:
+        case ForwardChatItemsErrorCode.noQuotedContent:
+        case ForwardChatItemsErrorCode.unknownForwardedDonation:
+        case ForwardChatItemsErrorCode.notEnoughFunds:
+        case ForwardChatItemsErrorCode.unallowedDonation:
         case ForwardChatItemsErrorCode.unknownUser:
         case ForwardChatItemsErrorCode.unknownForwardedAttachment:
         case ForwardChatItemsErrorCode.wrongItemsCount: