+++ CHANGELOG.md
@@ -6,6 +6,39 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
+## main
+
+[Diff](/../../compare/v0.7.0...main) | [Milestone](/../../milestone/60)
+
+### Fixed
+
+- UI:
+    - Media panel:
+        - Empty screen sharing being displayed sometimes. ([#1566])
+        - Incoming call window not being displayed in rare cases. ([#1567])
+        - Camera device turning off sometimes when microphone device is disconnected. ([#1568])
+
+[#1566]: /../../pull/1566
+[#1567]: /../../pull/1567
+[#1568]: /../../pull/1568
+
+
+
+
+## [0.7.0] · 2025-12-25
+[0.7.0]: /../../tree/v0.7.0
+
+[Diff](/../../compare/v0.6.14...v0.7.0) | [Milestone](/../../milestone/59)
+
+### Fixed
+
+- Authorization sometimes being lost on some devices. ([#1564])
+
+[#1564]: /../../pull/1564
+
+
+
+
 ## [0.6.14] · 2025-12-19
 [0.6.14]: /../../tree/v0.6.14